--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -40,7 +40,7 @@ jobs:
     - name: Setup Go
       uses: actions/setup-go@v3
       with:
-        go-version: "1.22.0"
+        go-version: "1.22"
     - name: checkout code
       uses: actions/checkout@v3
     - name: golangci-lint