--- a/perfmetrics/scripts/run_e2e_tests.sh
+++ b/perfmetrics/scripts/run_e2e_tests.sh
@@ -27,8 +27,8 @@ run_e2e_tests_on_package=$1
 
 # e.g. architecture=arm64 or amd64
 architecture=$(dpkg --print-architecture)
-echo "Installing go-lang 1.21.6..."
-wget -O go_tar.tar.gz https://go.dev/dl/go1.21.6.linux-${architecture}.tar.gz -q
+echo "Installing go-lang 1.22..."
+wget -O go_tar.tar.gz https://go.dev/dl/go1.22.linux-${architecture}.tar.gz -q
 sudo rm -rf /usr/local/go && tar -xzf go_tar.tar.gz && sudo mv go /usr/local
 export PATH=$PATH:/usr/local/go/bin
 # install python3-setuptools tools.