### Description
Upgrade to golang 1.22

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - Did sanity testing.
2. Unit tests - NA
3. Integration tests - Ran through presubmit - [passed](http://fusion2/ci/kokoro/prod%3Agcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/767422dc-ccd0-4c6c-a6ed-f34437a02db5)
4. Perf tests results
```

First run
---------

+--------+------------+--------------+------------+-------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW | RandWrite BW |
+--------+------------+--------------+------------+-------------+--------------+
| Master |  0.25MiB   |  496.7MiB/s  | 1.21MiB/s  |  71.83MiB/s |  1.27MiB/s   |
|   PR   |  0.25MiB   |  509.3MiB/s  |  1.3MiB/s  |  71.32MiB/s |  1.08MiB/s   |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
| Master | 48.828MiB  | 4754.43MiB/s | 77.42MiB/s | 256.34MiB/s |  47.12MiB/s  |
|   PR   | 48.828MiB  | 4722.55MiB/s | 76.01MiB/s | 210.91MiB/s |  44.0MiB/s   |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
| Master | 976.562MiB | 954.63MiB/s  | 3.28MiB/s  |  73.9MiB/s  |  3.18MiB/s   |
|   PR   | 976.562MiB | 575.55MiB/s  | 0.94MiB/s  |  38.65MiB/s |  1.41MiB/s   |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
+--------+------------+--------------+------------+-------------+--------------+

Second run
----------

+--------+------------+--------------+------------+-------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW | RandWrite BW |
+--------+------------+--------------+------------+-------------+--------------+
| Master |  0.25MiB   | 458.24MiB/s  | 1.21MiB/s  |  61.87MiB/s |  1.17MiB/s   |
|   PR   |  0.25MiB   | 509.53MiB/s  | 1.28MiB/s  |  73.72MiB/s |  1.36MiB/s   |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
| Master | 48.828MiB  | 744.32MiB/s  | 41.56MiB/s |  54.4MiB/s  |  41.57MiB/s  |
|   PR   | 48.828MiB  | 2348.94MiB/s | 56.97MiB/s | 127.66MiB/s |  38.79MiB/s  |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
| Master | 976.562MiB | 313.15MiB/s  |  0.7MiB/s  |  20.95MiB/s |  0.64MiB/s   |
|   PR   | 976.562MiB | 409.36MiB/s  | 0.78MiB/s  |  24.03MiB/s |  0.77MiB/s   |
|        |            |              |            |             |              |
|        |            |              |            |             |              |
+--------+------------+--------------+------------+-------------+--------------+

Third run
---------

+--------+------------+-------------+------------+-------------+--------------+
| Branch | File Size  |   Read BW   |  Write BW  | RandRead BW | RandWrite BW |
+--------+------------+-------------+------------+-------------+--------------+
| Master |  0.25MiB   | 469.33MiB/s | 1.16MiB/s  |  63.59MiB/s |  1.18MiB/s   |
|   PR   |  0.25MiB   | 465.15MiB/s | 1.27MiB/s  |  55.75MiB/s |  1.18MiB/s   |
|        |            |             |            |             |              |
|        |            |             |            |             |              |
| Master | 48.828MiB  | 774.08MiB/s | 44.89MiB/s |  71.5MiB/s  |  40.31MiB/s  |
|   PR   | 48.828MiB  | 536.99MiB/s | 40.58MiB/s |  57.61MiB/s |  40.58MiB/s  |
|        |            |             |            |             |              |
|        |            |             |            |             |              |
| Master | 976.562MiB |  326.9MiB/s | 0.67MiB/s  |  27.56MiB/s |  0.78MiB/s   |
|   PR   | 976.562MiB |  624.6MiB/s | 1.96MiB/s  |  45.29MiB/s |  1.98MiB/s   |
|        |            |             |            |             |              |
|        |            |             |            |             |              |
+--------+------------+-------------+------------+-------------+--------------+

```